Old operation_schema interface
